# Cache the reek configuration loading for faster execution