@etagwerker 2025-04-09
@raff-s Yes, it would. Here it is: 

- https://github.com/whitesmith/rubycritic/blob/main/CHANGELOG.md#v492--2025-04-08-commits 
- https://rubygems.org/gems/rubycritic